refactor(server): extract /api/email-campaign/* into a router module
Second route-group extraction. Moves all 9 email-campaign handlers out
of server.js into src/server/routes/email-campaign.js, mounted via
`app.use('/api/email-campaign', requireAuth, emailCampaignRouter)`.

Routes (paths byte-identical, verified by the route guard):
  POST /api/email-campaign/create
  GET  /api/email-campaign/list/:brandProfileId
  GET  /api/email-campaign/:id
  PATCH /api/email-campaign/email/:id
  POST /api/email-campaign/email/:id/resolve-flag
  POST /api/email-campaign/email/:id/dismiss-flag-as-false-positive
  GET  /api/email-campaign/generate/:id
  POST /api/email-campaign/save-brief-template
  GET  /api/email-campaign/brief-templates/:brandProfileId

Auth: requireAuth at the mount (all 9 were uniformly authed) — same
convention as the compliance router.

Also extracts a shared dependency the no-undef gate surfaced:
- src/server/streams.js — the globalThis-backed `activeStreams` SSE
  registry. The email-campaign generate handler shares it with the
  social-generator generate handler (still inline in server.js), so it
  had to become a shared module rather than move into either router.
  Both server.js and the email-campaign module now import it; social-gen
  will import it when it's extracted.

Module imports: express, fs, path, pool (db), safeParseLLM (llm-json),
stripScaffoldingArtifacts (text), dateContext (llm), activeStreams
(streams). Handler bodies verbatim; only registration lines changed.

Test plan:
- node --check server.js + email-campaign.js + streams.js -> OK
- npm run lint (no-undef) -> clean (it caught activeStreams + fs/path on
  the first pass — fixed by the streams.js extraction + module imports)
- route-inventory guard -> 213, byte-identical (9 paths reconstructed)
- vitest -> 93/93

Rollback: revert; verbatim move + a 6-line shared module.
